Update UdpServer.cpp
Needed to add memset statement as first statement in while loop to zero out buffer prior to reading in new data. This fixes a bug whereby if a long message is followed by a shorter message on Ubuntu under WSL2 on Windows that then the message read from the socket is not "\0" terminated and therefore the end of the previous message remains in the buffer - this causes SIP transactions to fail in the clients.